requirements.txt

customtkinter==5.2.0
openpyxl==3.1.2
pandas==2.1.0
Pillow==10.0.1
python-dotenv==1.0.0
selenium==4.12.0
webdriver-manager==4.0.0
xlrd==2.0.1
